size_t for messageSizes

+++ src/Massage/MassagePacker.h
@@ -116,7 +116,7 @@ const char* buffer  () const {
     }
 
       // Current size of message in buffer.
-    uint8_t _messageSize;
+    size_t _messageSize;
 
     // Index in the buffer of next argument to read.
     uint8_t _nextIndex;
